Rearrange `context` module.
Separate out top level modules for file abstractions and for filesystem
backed data store.
