+++ docs/python.rst
@@ -214,7 +214,7 @@ Base classes
         loop.run()
 
 .. I list the members explicitly because nothing else seems to suppress the documentation of __fspath__ in Sphinx 4.1.2
-.. autoclass:: scalems.context._file.FileReference
+.. autoclass:: scalems.file.FileReference
     :members: is_local, path, as_uri, localize
 
 Logging
@@ -228,4 +228,4 @@ Exceptions
 .. automodule:: scalems.exceptions
     :members:
 
-.. autoclass:: scalems.context._file.DataLocalizationError
+.. autoclass:: scalems.file.DataLocalizationError
