+++ src/scalems/exceptions.py
@@ -39,6 +39,10 @@ class APIError(ScaleMSError):
     """Specified interfaces are being violated."""
 
 
+class ContextError(ScaleMSError):
+    """A Context operation could not be performed."""
+
+
 class DispatchError(ScaleMSError):
     """SCALE-MS is unable to execute work or manage data in the requested environment."""
