+++ src/scalems/store/_lock.py
