+++ tests/test_context_datastore.py
@@ -7,10 +7,9 @@ import tempfile
 
 import pytest
 import scalems.context as _context
-import scalems.context._lock
 import scalems.exceptions
-from scalems.context._datastore import filestore_generator
-from scalems.context import FileStore, FileStoreManager
+from scalems.store import filestore_generator
+from scalems.store import FileStore, FileStoreManager
 
 
 def test_normal_lifecycle(tmp_path, caplog):
@@ -33,12 +32,12 @@ def test_normal_lifecycle(tmp_path, caplog):
 
     generator = filestore_generator(directory=tmp_path)
     datastore = next(generator)()
-    with pytest.raises(scalems.context.ContextError):
+    with pytest.raises(scalems.exceptions.ContextError):
         FileStore(directory=tmp_path)
     datastore.close()
     # finalize_datastore() must be called exactly once for a data store that has been
     # opened.
-    with pytest.raises(scalems.context.StaleFileStore):
+    with pytest.raises(scalems.store.StaleFileStore):
         datastore.close()
     with pytest.raises(StopIteration):
         with pytest.warns(scalems.exceptions.ScaleMSWarning):
@@ -46,12 +45,10 @@ def test_normal_lifecycle(tmp_path, caplog):
     del generator
 
     # Confirm some assumptions about implementation details.
-    filepath = tmp_path.joinpath(
-        scalems.context._datastore._data_subdirectory, scalems.context._datastore._metadata_filename
-    )
+    filepath = tmp_path.joinpath(scalems.store._data_subdirectory, scalems.store._metadata_filename)
 
     manager = FileStoreManager(tmp_path)
-    datastore: scalems.context.FileStore = manager.filestore()
+    datastore: scalems.store.FileStore = manager.filestore()
     assert datastore.instance == os.getpid()
     assert datastore.filepath == filepath
     with pytest.raises(AttributeError):
@@ -59,7 +56,7 @@ def test_normal_lifecycle(tmp_path, caplog):
         datastore.log = list()
     # TODO: log interface.
     # datastore.log.append('Testing')
-    with caplog.at_level(logging.CRITICAL, logger="scalems.context._datastore"):
+    with caplog.at_level(logging.CRITICAL, logger="scalems.store"):
         assert not datastore.closed
         del manager
         assert datastore.closed
@@ -83,22 +80,22 @@ def test_nonfinalized(tmp_path, caplog):
         # protections aren't violated.
         # Right now, we have to manipulate the filesystem with knowledge of the implementation
         # details in order to effectively test.
-        with pytest.raises(scalems.context.ContextError):
+        with pytest.raises(scalems.exceptions.ContextError):
             FileStoreManager().filestore()
         # We may want to assert constraints on the filesystem state we expect to encounter
         # even when a lock is left unexpectedly, but initially all we know is that a
         # dangling lock may result from an unclean process termination.
-        scalems.context._lock._lock_directory()
-        with pytest.raises(scalems.context.ContextError):
+        scalems.store._lock._lock_directory()
+        with pytest.raises(scalems.exceptions.ContextError):
             FileStoreManager()
-        scalems.context._lock._unlock_directory()
+        scalems.store._lock._unlock_directory()
 
         logger = logging.getLogger("scalems.context._datastore")
         logger.debug(
             "Suppressing an error that would normally occur when we fail to "
             "close the FileStore (that we intentionally broke in this test)."
         )
-        with caplog.at_level(logging.CRITICAL, logger="scalems.context._datastore"):
+        with caplog.at_level(logging.CRITICAL, logger="scalems.store"):
             del manager
 
 
@@ -111,15 +108,15 @@ def test_contention(tmp_path, caplog):
     with _context.scoped_chdir(tmp_path):
         manager = FileStoreManager()
         datastore = manager.filestore()
-        scalems.context._lock._lock_directory()
-        with pytest.raises(scalems.context._lock.LockException):
+        scalems.store._lock._lock_directory()
+        with pytest.raises(scalems.store._lock.LockException):
             datastore.close()
         # This failed call to `close` leaves the datastore in an invalid state.
-        with pytest.raises(scalems.context._datastore.ContextError):
+        with pytest.raises(scalems.exceptions.ContextError):
             assert not datastore.closed
 
-        scalems.context._lock._unlock_directory()
-        with caplog.at_level(logging.CRITICAL, logger="scalems.context._datastore"):
+        scalems.store._lock._unlock_directory()
+        with caplog.at_level(logging.CRITICAL, logger="scalems.store"):
             # The above failed `close` left the manager in an invalid state.
             del manager
 
@@ -128,13 +125,13 @@ def test_contention(tmp_path, caplog):
         unexpected_instance = expected_instance + 1
         with open(metadata_path, "w") as fp:
             json.dump({"instance": unexpected_instance}, fp)
-        with pytest.raises(scalems.context.ContextError):
+        with pytest.raises(scalems.exceptions.ContextError):
             FileStoreManager()
 
-        scalems.context._lock._lock_directory()
-        with pytest.raises(scalems.context._datastore.ContextError):
+        scalems.store._lock._lock_directory()
+        with pytest.raises(scalems.exceptions.ContextError):
             FileStoreManager()
-        scalems.context._lock._unlock_directory()
+        scalems.store._lock._unlock_directory()
 
 
 def test_nesting(tmp_path):
@@ -144,10 +141,10 @@ def test_nesting(tmp_path):
     structures.
     """
     # initialization should check for locks in parent directories.
-    with _context._lock.scoped_directory_lock(tmp_path):
+    with scalems.store._lock.scoped_directory_lock(tmp_path):
         with tempfile.TemporaryDirectory(dir=tmp_path) as path:
             with _context.scoped_chdir(path):
-                with pytest.raises(scalems.context._datastore.ContextError):
+                with pytest.raises(scalems.exceptions.ContextError):
                     FileStoreManager(directory=path)
 
     # initialization should check for unexpected nesting in the work dir.
@@ -158,12 +155,12 @@ def test_nesting(tmp_path):
     datastore_path = datastore.datastore
     with tempfile.TemporaryDirectory(dir=root_path) as path:
         with _context.scoped_chdir(path):
-            with pytest.raises(scalems.context._datastore.ContextError):
+            with pytest.raises(scalems.exceptions.ContextError):
                 FileStoreManager()
 
     # initialization should check for unexpected nesting in the datastore itself.
     with _context.scoped_chdir(datastore_path):
-        with pytest.raises(scalems.context._datastore.ContextError):
+        with pytest.raises(scalems.exceptions.ContextError):
             FileStoreManager()
